JSON

+++ src/main/java/Figures/Trasform/IScaleble.java
@@ -1,5 +1,5 @@
 package Figures.Trasform;
 
 public interface IScaleble{
-    public void Scale(int x);
+    public void Scale(double x);
 }
